+++ _tasks/downloads.ts
@@ -2,9 +2,9 @@ import { chain as kusama } from "kusama/mod.ts"
 import { chain as polkadot } from "polkadot/mod.ts"
 import { chain as rococo } from "rococo/mod.ts"
 import { chain as westend } from "westend/mod.ts"
+import { hex } from "../crypto/mod.ts"
 import { emptyDir } from "../deps/std/fs.ts"
 import dprintConfig from "../dprint.json" assert { type: "json" }
-import { hex } from "../util/mod.ts"
 
 const GENERATION_NOTICE = `// This file was generated by \`_tasks/downloads.ts\``
 
@@ -18,11 +18,11 @@ await Deno.writeTextFile(
   `${GENERATION_NOTICE}
 import { createFromBuffer } from "../deps/dprint.ts"
 import dprintConfig from "../dprint.json" assert { type: "json" }
-import { decode } from "./hex.ts"
+import { hex } from "./hex.ts"
 
 const wasm = "${wasm}"
 
-export const tsFormatter = createFromBuffer(decode(wasm))
+export const tsFormatter = createFromBuffer(hex.decode(wasm))
 const { indentWidth, lineWidth, typescript: config } = dprintConfig
 tsFormatter.setConfig({ indentWidth, lineWidth }, config)
 `,
