+++ crypto/mod.ts
@@ -1,0 +1,8 @@
+export * as hashers from "./hashers.ts"
+export * as hex from "./hex.ts"
+export * as ss58 from "./ss58.ts"
+
+// moderate --exclude hashers.ts hex.ts ss58.ts
+
+export * from "./Sr25519.ts"
+export * from "./test_pairs.ts"
