+++ frame_metadata/_downloaded/mod.ts
@@ -1,5 +1,5 @@
 // This file was generated by `_tasks/downloads.ts`
-import { hex } from "../../util/mod.ts"
+import { hex } from "../../crypto/mod.ts"
 import { $metadata } from "../mod.ts"
 
 export const [
