+++ mod.ts
@@ -1,27 +1,10 @@
 export * as $ from "./deps/scale.ts"
 export { BitSequence } from "./deps/scale.ts"
 export * as frame from "./frame_metadata/mod.ts"
-export {
-  alice,
-  aliceStash,
-  bob,
-  bobStash,
-  charlie,
-  dave,
-  End,
-  eve,
-  ferdie,
-  getOrInit,
-  type Hex,
-  hex,
-  type HexHash,
-  type Listener,
-  Sr25519,
-  ss58,
-} from "./util/mod.ts"
 
 // moderate --exclude frame_metadata main.ts providers server util
 
+export * from "./crypto/mod.ts"
 export * from "./fluent/mod.ts"
 export * from "./primitives/mod.ts"
 export * from "./rpc/mod.ts"
