+++ patterns/consensus/babeBlockAuthor.ts
@@ -2,12 +2,11 @@ import { $preDigest } from "polkadot_dev/types/sp_consensus_babe/digests.ts"
 import { Chain, ChainRune } from "../../fluent/mod.ts"
 import { PublicKeyRune } from "../../fluent/mod.ts"
 import { Rune, RunicArgs, ValueRune } from "../../rune/mod.ts"
-import { HexHash } from "../../util/branded.ts"
 import { preRuntimeDigest } from "./preRuntimeDigest.ts"
 
 export function babeBlockAuthor<U, C extends Chain, X>(
   chain: ChainRune<U, C>,
-  ...[blockHash]: RunicArgs<X, [HexHash]>
+  ...[blockHash]: RunicArgs<X, [string]>
 ) {
   const validators = chain
     .metadata()
