+++ rpc/known/beefy.ts
@@ -1,4 +1,4 @@
-import { Hash, Hex, Subscription } from "./utils.ts"
+import { Subscription } from "./utils.ts"
 
 // https://github.com/paritytech/substrate/blob/317808a/client/beefy/rpc/src/lib.rs#L84
 export type BeefyCalls = {
@@ -9,10 +9,10 @@ export type BeefyCalls = {
    * in the network or if the client is still initializing or syncing with the network.
    * In such case an error would be returned.
    */
-  beefy_getFinalizedHead(): Hash
+  beefy_getFinalizedHead(): string
 }
 
 export type BeefySubscriptions = {
   /** Returns the block most recently finalized by BEEFY, alongside side its justification. */
-  beefy_subscribeJustifications(): Subscription<"beefy_unsubscribeJustifications", Hex>
+  beefy_subscribeJustifications(): Subscription<"beefy_unsubscribeJustifications", string>
 }
