+++ rpc/known/childstate.ts
@@ -1,16 +1,14 @@
-import { Hash, Hex } from "./utils.ts"
-
 // https://github.com/paritytech/substrate/blob/4d04aba/primitives/storage/src/lib.rs
-export type StorageKey = Hex
-export type PrefixedStorageKey = Hex
-export type StorageData = Hex
+export type StorageKey = string
+export type PrefixedStorageKey = string
+export type StorageData = string
 
 // https://github.com/paritytech/substrate/blob/ded44948/client/rpc-api/src/state/helpers.rs#L27
 export interface ReadProof {
   /** Block hash used to generate the proof */
-  at: Hash
+  at: string
   /** A proof used to prove that storage entries are included in the storage trie */
-  proof: Hex[]
+  proof: string[]
 }
 
 // https://github.com/paritytech/substrate/blob/934fbfd/client/rpc-api/src/child_state/mod.rs#L29
@@ -22,7 +20,7 @@ export type ChildStateCalls = {
   childState_getKeys(
     childStorageKey: PrefixedStorageKey,
     prefix: StorageKey,
-    hash?: Hash,
+    hash?: string,
   ): StorageKey[]
   /**
    * Returns the keys with prefix from a child storage with pagination support.
@@ -34,36 +32,36 @@ export type ChildStateCalls = {
     prefix: StorageKey,
     count: number,
     startKey?: StorageKey,
-    hash?: Hash,
+    hash?: string,
   ): StorageKey[]
   /** Returns a child storage entry at a specific block's state. */
   childState_getStorage(
     childStorageKey: PrefixedStorageKey,
     key: StorageKey,
-    hash?: Hash,
+    hash?: string,
   ): StorageData | null
   /** Returns child storage entries for multiple keys at a specific block's state. */
   childState_getStorageEntries(
     childStorageKey: PrefixedStorageKey,
     keys: StorageKey[],
-    hash?: Hash,
+    hash?: string,
   ): (StorageData | null)[]
   /** Returns the hash of a child storage entry at a block's state. */
   childState_getStorageHash(
     childStorageKey: PrefixedStorageKey,
     key: StorageKey,
-    hash?: Hash,
-  ): Hash | null
+    hash?: string,
+  ): string | null
   /** Returns the size of a child storage entry at a block's state. */
   childState_getStorageSize(
     childStorageKey: PrefixedStorageKey,
     key: StorageKey,
-    hash?: Hash,
+    hash?: string,
   ): number | null
   /** Returns proof of storage for child key entries at a specific block's state. */
   state_getChildReadProof(
     childStorageKey: PrefixedStorageKey,
     keys: StorageKey[],
-    hash?: Hash,
+    hash?: string,
   ): ReadProof
 }
