+++ rpc/known/smoldot.ts
@@ -1,4 +1,4 @@
-import { Hash, Hex, Subscription } from "./utils.ts"
+import { Subscription } from "./utils.ts"
 
 export type NetworkConfig = {
   totalAttempts: number
@@ -20,25 +20,25 @@ export type RuntimeSpec = {
   specVersion: number
   implVersion: number
   transactionVersion?: number
-  apis: Record<Hex, number>
+  apis: Record<string, number>
 }
 
 export type ChainHeadUnstableFollowEvent = {
   event: "initialized"
-  finalizedBlockHash: Hash
+  finalizedBlockHash: string
   finalizedBlockRuntime?: MaybeRuntimeSpec
 } | {
   event: "newBlock"
-  blockHash: Hash
-  parentBlockHash: Hash
+  blockHash: string
+  parentBlockHash: string
   newRuntime?: MaybeRuntimeSpec
 } | {
   event: "bestBlockChanged"
-  bestBlockHash: Hash
+  bestBlockHash: string
 } | {
   event: "finalized"
-  finalizedBlockHashes: Hash[]
-  prunedBlockHashes: Hash[]
+  finalizedBlockHashes: string[]
+  prunedBlockHashes: string[]
 } | {
   event: "stop"
 }
@@ -67,47 +67,47 @@ export type TransactionWatchEvent = {
 }
 
 export type TransactionWatchEventBlock = {
-  hash: Hash
+  hash: string
   index: string
 }
 
 export type SmoldotRpc = {
   chainHead_unstable_body(
     followSubscription: string,
-    hash: Hash,
+    hash: string,
     networkConfig?: NetworkConfig,
   ): string
   chainHead_unstable_call(
     followSubscription: string,
-    hash: Hash,
+    hash: string,
     fn: string,
-    callParameters: Hex,
+    callParameters: string,
     networkConfig?: NetworkConfig,
   ): string
   chainHead_unstable_follow(
     runtimeUpdates: boolean,
   ): Subscription<"chainHead_unstable_follow", ChainHeadUnstableFollowEvent>
-  chainHead_unstable_genesisHash(): Hash
-  chainHead_unstable_header(followSubscription: string, hash: Hash): Hex
+  chainHead_unstable_genesisHash(): string
+  chainHead_unstable_header(followSubscription: string, hash: string): string
   chainHead_unstable_stopBody(subscription: string): void
   chainHead_unstable_stopCall(subscription: string): void
   chainHead_unstable_stopStorage(subscription: string): void
   chainHead_unstable_storage(
     followSubscription: string,
-    hash: Hash,
-    key: Hex,
-    childKey?: Hex,
+    hash: string,
+    key: string,
+    childKey?: string,
     networkConfig?: NetworkConfig,
   ): string
   chainHead_unstable_unfollow(followSubscription: string): void
-  chainHead_unstable_unpin(followSubscription: string, hash: Hash): void
+  chainHead_unstable_unpin(followSubscription: string, hash: string): void
   chainSpec_unstable_chainName(): string
-  chainSpec_unstable_genesisHash(): Hash
+  chainSpec_unstable_genesisHash(): string
   chainSpec_unstable_properties(): Record<string, unknown>
   sudo_unstable_p2pDiscover(multiaddr: string): void
   sudo_unstable_version(): string
   transaction_unstable_submitAndWatch(
-    transaction: Hex,
+    transaction: string,
   ): Subscription<"transaction_unstable_submitAndWatch", TransactionWatchEvent>
   transaction_unstable_unwatch(subscription: string): void
   chainHead_unstable_finalizedDatabase(maxSizeBytes?: bigint): string
