--- util/branded.ts
@@ -1,4 +1,0 @@
-// TODO: delete these upon re-rpc merge
-export type Hex = string
-export type Hash = string
-export type HexHash = string
