--- util/types.ts
@@ -1,1 +1,0 @@
-export type PromiseOr<T> = T | Promise<T>
